+++ src/ultrasphere/integral.py
@@ -1,0 +1,196 @@
+from ultrasphere.coordinates import BranchingType, SphericalCoordinates, get_child
+
+
+from collections.abc import Mapping
+
+
+def roots(
+    c: SphericalCoordinates[TSpherical, TEuclidean],
+    n: int,
+    *,
+    expand_dims_x: bool,
+    expand_dims_w: bool = False,
+    device: ivy.Device | ivy.NativeDevice | None = None,
+    dtype: ivy.Dtype | ivy.NativeDtype | None = None,
+) -> tuple[Mapping[TSpherical, Array], Mapping[TSpherical, Array]]:
+    """
+    Gauss-Jacobi quadrature roots and weights.
+
+    Parameters
+    ----------
+    n : int
+        The number of roots.
+    expand_dims_x : bool
+        Whether to expand dimensions of the roots, by default False
+    expand_dims_w : bool, optional
+        Whether to expand dimensions of the weights, by default False
+    device : ivy.Device | ivy.NativeDevice, optional
+        The device, by default None
+    dtype : ivy.Dtype | ivy.NativeDtype, optional
+        The data type, by default None
+
+    Returns
+    -------
+    tuple[Mapping[TSpherical, Array], Mapping[TSpherical, Array]]
+        roots and weights
+
+    Raises
+    ------
+    ValueError
+        If the branching type is invalid.
+
+    """
+    from scipy.special import roots_jacobi
+
+    xs = {}
+    ws = {}
+    for i, node in enumerate(c.s_nodes):
+        branching_type = c.branching_types[node]
+        if branching_type == BranchingType.A:
+            x = ivy.arange(2 * n, device=device, dtype=dtype) * ivy.pi / n
+            w = ivy.ones(2 * n, device=device, dtype=dtype) * ivy.pi / n
+        elif branching_type == BranchingType.B:
+            s_beta = c.S[get_child(c.G, node, "sin")]
+            beta = s_beta / 2
+            x, w = roots_jacobi(n, beta, beta)
+            x = ivy.acos(x)
+        elif branching_type == BranchingType.BP:
+            s_alpha = c.S[get_child(c.G, node, "cos")]
+            alpha = s_alpha / 2
+            x, w = roots_jacobi(n, alpha, alpha)
+            x = ivy.asin(x)
+        elif branching_type == BranchingType.C:
+            s_alpha = c.S[get_child(c.G, node, "cos")]
+            s_beta = c.S[get_child(c.G, node, "sin")]
+            alpha = s_alpha / 2
+            beta = s_beta / 2
+            x, w = roots_jacobi(n, alpha, beta)
+            w /= 2 ** (alpha + beta + 2)
+            x = ivy.acos(x) / 2
+        else:
+            raise ValueError(f"Invalid branching type {branching_type}.")
+        x = ivy.array(x, device=device, dtype=dtype)
+        w = ivy.array(w, device=device, dtype=dtype)
+        if expand_dims_x:
+            x = x[(None,) * i + (slice(None),) + (None,) * (c.s_ndim - i - 1)]
+        if expand_dims_w:
+            w = w[(None,) * i + (slice(None),) + (None,) * (c.s_ndim - i - 1)]
+        xs[node] = x
+        ws[node] = w
+    return xs, ws
+
+
+@overload
+def integrate(
+    c: SphericalCoordinates[TSpherical, TEuclidean],
+    f: (
+        Callable[
+            [Mapping[TSpherical, Array | NativeArray]],
+            Mapping[TSpherical, Array | NativeArray],
+        ]
+        | Mapping[TSpherical, Array | NativeArray]
+    ),
+    does_f_support_separation_of_variables: Literal[True],
+    n: int,
+    *,
+    device: ivy.Device | ivy.NativeDevice | None = None,
+    dtype: ivy.Dtype | ivy.NativeDtype | None = None,
+) -> Mapping[TSpherical, Array]: ...
+
+@overload
+def integrate(
+    c: SphericalCoordinates[TSpherical, TEuclidean],
+    f: (
+        Callable[
+            [Mapping[TSpherical, Array | NativeArray]],
+            Array | NativeArray,
+        ]
+        | Array
+        | NativeArray
+    ),
+    does_f_support_separation_of_variables: Literal[False],
+    n: int,
+    *,
+    device: ivy.Device | ivy.NativeDevice | None = None,
+    dtype: ivy.Dtype | ivy.NativeDtype | None = None,
+) -> Array: ...
+
+def integrate(
+    c: SphericalCoordinates[TSpherical, TEuclidean],
+    f: (
+        Callable[
+            [Mapping[TSpherical, Array | NativeArray]],
+            Mapping[TSpherical, Array | NativeArray] | Array | NativeArray,
+        ]
+        | Mapping[TSpherical, Array | NativeArray]
+        | Array
+        | NativeArray
+    ),
+    does_f_support_separation_of_variables: bool,
+    n: int,
+    *,
+    device: ivy.Device | ivy.NativeDevice | None = None,
+    dtype: ivy.Dtype | ivy.NativeDtype | None = None,
+) -> Array | Mapping[TSpherical, Array]:
+    """
+    Integrate the function over the hypersphere.
+
+    Parameters
+    ----------
+    f : Callable[ [Mapping[TSpherical, Array  |  NativeArray]],
+        Mapping[TSpherical, Array  |  NativeArray]  |  Array  |  NativeArray, ]
+        |  Mapping[TSpherical, Array  |  NativeArray]  |  Array  |  NativeArray
+            The function to integrate or the values of the function.
+        In case of vectorized function, the function should add extra
+        axis to the last dimension, not the first dimension.
+    does_f_support_separation_of_variables : bool
+        Whether the function supports separation of variables.
+        This could significantly reduce the computational cost.
+    n : int
+        The number of roots.
+    device : ivy.Device | ivy.NativeDevice, optional
+        The device, by default None
+    dtype : ivy.Dtype | ivy.NativeDtype, optional
+        The data type, by default None
+
+    Returns
+    -------
+    Array | Mapping[TSpherical, Array]
+        The integrated value.
+
+    """
+    xs, ws = c.roots(
+        n,
+        device=device,
+        dtype=dtype,
+        expand_dims_x=not does_f_support_separation_of_variables,
+    )
+    if isinstance(f, Callable):  # type: ignore
+        try:
+            val = f(xs)  # type: ignore
+        except Exception as e:
+            raise RuntimeError(f"Error occurred while evaluating {f=}") from e
+    else:
+        val = f
+
+    # in case f(theta1, ...) = f_1(theta1) * f_2(theta2) * ...
+    if isinstance(val, Mapping):
+        result = {}
+        for node in c.s_nodes:
+            value = val[node]
+            # supports vectorized function
+            # axis=0 because in sph_harm
+            # we add axis to the last dimension
+            # theta(node),u1,...,uM
+            ivy.broadcast_shapes(ivy.shape(value)[:1], ivy.shape(ws[node]))
+            w = ws[node].reshape([-1] + [1] * (value.ndim - 1))
+            result[node] = ivy.sum(value * w, axis=0)
+        # we don't know how to einsum the result
+        return result
+    # theta1,...,thetaN,u1,...,uM\
+    for node in c.s_nodes:
+        w = ws[node]
+        ivy.broadcast_shapes(ivy.shape(val)[:1], ivy.shape(w))
+        # val = ivy.einsum("i...,i->...", val, w.astype(val.dtype))
+        val = ivy.sum(val * w[(slice(None),) + (None,) * (val.ndim - 1)], axis=0)
+    return val
